shoe_detection_websocket/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
blinker==1.8.2
cffi==1.17.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
protobuf==3.20.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.44.0
sympy==1.13.3
torch==2.2.2
typing_extensions==4.13.1
uvicorn==0.33.0
websockets==13.1
Werkzeug==3.0.6
zipp==3.20.2
